% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{choose_def_coverage}
\alias{choose_def_coverage}
\title{Choose defensive coverage}
\usage{
choose_def_coverage(
  down,
  togo,
  YdsBef,
  posstmdiff,
  quarter_secs,
  quarter,
  personnel,
  def_personnel,
  formation
)
}
\arguments{
\item{down}{Current down.}

\item{togo}{Yards to go.}

\item{YdsBef}{Yards to end zone before the play.}

\item{posstmdiff}{Score differential for the possession team.}

\item{quarter_secs}{Seconds remaining in the quarter.}

\item{quarter}{Current quarter.}

\item{personnel}{The offensive personnel on field}

\item{def_personnel}{The defensive personnel on field}

\item{formation}{The offensive formation on field}
}
\value{
A character string.
}
\description{
Chooses defensive coverage for a simulated play.
}
