% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{choose_def_personnel}
\alias{choose_def_personnel}
\title{Choose defensive personnel}
\usage{
choose_def_personnel(
  down,
  togo,
  YdsBef,
  posstmdiff,
  quarter_secs,
  quarter,
  personnel
)
}
\arguments{
\item{down}{Current down.}

\item{togo}{Yards to go.}

\item{YdsBef}{Yards to end zone before the play.}

\item{posstmdiff}{Score differential for the possession team.}

\item{quarter_secs}{Seconds remaining in the quarter.}

\item{quarter}{Current quarter.}

\item{personnel}{The offensive personnel on field}
}
\value{
A character string.
}
\description{
Chooses defensive personnel for a simulated play.
}
