% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{choose_formation}
\alias{choose_formation}
\title{Choose offensive formation}
\usage{
choose_formation(
  down,
  togo,
  YdsBef,
  posstmdiff,
  quarter_secs,
  quarter,
  personnel,
  def_personnel
)
}
\arguments{
\item{down}{Current down.}

\item{togo}{Yards to go.}

\item{YdsBef}{Yards to end zone before the play.}

\item{posstmdiff}{Score differential for the possession team.}

\item{quarter_secs}{Seconds remaining in the quarter.}

\item{quarter}{Current quarter.}

\item{personnel}{The offensive personnel on field}

\item{def_personnel}{The defensive personnel on field}
}
\value{
A character string.
}
\description{
Chooses offensive formation for a simulated play.
}
