% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{choose_personnel}
\alias{choose_personnel}
\title{Choose offensive personnel}
\usage{
choose_personnel(down, togo, YdsBef, posstmdiff, quarter_secs, quarter)
}
\arguments{
\item{down}{Current down.}

\item{togo}{Yards to go.}

\item{YdsBef}{Yards to end zone before the play.}

\item{posstmdiff}{Score differential for the possession team.}

\item{quarter_secs}{Seconds remaining in the quarter.}

\item{quarter}{Current quarter.}
}
\value{
A character string.
}
\description{
Chooses offensive personnel for a simulated play.
}
