% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{full_PFF}
\alias{full_PFF}
\title{Creates a nice looking data frame of PFF data}
\usage{
full_PFF(tm, year)
}
\arguments{
\item{tm}{Character string. A two- or three-letter NFL team abbreviation.
Must be one of the following:
\code{"ARI"}, \code{"ATL"}, \code{"BAL"}, \code{"BUF"}, \code{"CAR"},
\code{"CHI"}, \code{"CIN"}, \code{"CLE"}, \code{"DAL"}, \code{"DEN"},
\code{"DET"}, \code{"GB"}, \code{"HOU"}, \code{"IND"}, \code{"JAX"},
\code{"KC"}, \code{"LA"}, \code{"LAC"}, \code{"LV"}, \code{"MIA"},
\code{"MIN"}, \code{"NE"}, \code{"NO"}, \code{"NYG"}, \code{"NYJ"},
\code{"PHI"}, \code{"PIT"}, \code{"SEA"}, \code{"SF"}, \code{"TB"},
\code{"TEN"}, \code{"WAS"}.}

\item{year}{A year between 2024 and 2025}
}
\value{
A \code{data.frame} representing a nice looking data frame of
PFF data including mean and sd PFF grades for each position
}
\description{
Creates a nice looking data frame of PFF data
}
