% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{multiple_simulations}
\alias{multiple_simulations}
\title{Give multiple simulations}
\usage{
multiple_simulations(team1, team2, year = 2025, n = 100, max_attempts = 3)
}
\arguments{
\item{team1}{An NFL team}

\item{team2}{A different NFL team}

\item{year}{A year between 2024 and 2025}

\item{n}{Number of simulations desired}

\item{max_attempts}{Number of failed attempts allowed}
}
\value{
A data frame.
}
\description{
Returns n simulated games for a given matchup and year.
}
