% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{routeselection1}
\alias{routeselection1}
\title{Give play details}
\usage{
routeselection1(
  posstm,
  deftm,
  down,
  togo,
  YdsBef,
  posstmdiff,
  quarter_secs,
  quarter,
  off_dat,
  def_dat,
  year
)
}
\arguments{
\item{posstm}{Team with possession of the ball.}

\item{deftm}{Team on defense}

\item{down}{Current down.}

\item{togo}{Yards to go.}

\item{YdsBef}{Yards to end zone before the play.}

\item{posstmdiff}{Score differential for the possession team.}

\item{quarter_secs}{Seconds remaining in the quarter.}

\item{quarter}{Current quarter.}

\item{off_dat}{The PFF data frame for the offensive team}

\item{def_dat}{The PFF data frame for the defensive team}

\item{year}{A year between 2024 and 2025}
}
\value{
A named list.
}
\description{
Returns play details for a simulated play.
}
