% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{simulator}
\alias{simulator}
\title{Simulates an NFL football game}
\usage{
simulator(team1, team2, year = 2025, track = "NO")
}
\arguments{
\item{team1}{An NFL team}

\item{team2}{A different NFL team}

\item{year}{A year between 2024 and 2025}

\item{track}{Allows the user to decide whether to track the game or not}
}
\value{
A data frame.
}
\description{
Returns a simulated game for a provided matchup and year.
}
