% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{runorpass}
\alias{runorpass}
\title{Choose run or pass}
\usage{
runorpass(
  down,
  togo,
  YdsBef,
  posstmdiff,
  quarter_secs,
  quarter,
  personnel,
  formation,
  def_personnel,
  coverage
)
}
\arguments{
\item{down}{Current down.}

\item{togo}{Yards to go.}

\item{YdsBef}{Yards to end zone before the play.}

\item{posstmdiff}{Score differential for the possession team.}

\item{quarter_secs}{Seconds remaining in the quarter.}

\item{quarter}{Current quarter.}

\item{personnel}{The offensive personnel on field}

\item{formation}{The offensive formation on field}

\item{def_personnel}{The defensive personnel on field}

\item{coverage}{The defensive coverage on field}
}
\value{
A character string.
}
\description{
Chooses run or pass for a simulated play.
}
